+++ Sources/KeyboardKit/Actions/KeyboardAction+Autocomplete.swift
@@ -16,10 +16,10 @@ public extension KeyboardAction {
      */
     var shouldApplyAutocorrectSuggestion: Bool {
         switch self {
-        case .character(let char): return char.isWordDelimiter
-        case .primary(let type): return type.isSystemAction
-        case .space: return true
-        default: return false
+        case .character(let char): char.isWordDelimiter
+        case .primary(let type): type.isSystemAction
+        case .space: true
+        default: false
         }
     }
     
@@ -37,8 +37,8 @@ public extension KeyboardAction {
      */
     var shouldRemoveAutocompleteInsertedSpace: Bool {
         switch self {
-        case .character(let char): return char.isWordDelimiter && self != .space
-        default: return false
+        case .character(let char): char.isWordDelimiter && self != .space
+        default: false
         }
     }
 }