+++ Sources/KeyboardKit/_Keyboard/Keyboard+Case.swift
@@ -45,3 +45,20 @@ public extension Keyboard.Case {
         }
     }
 }
+
+extension Keyboard.Case {
+    
+    var standardPressAction: KeyboardAction.GestureAction? {
+        switch self {
+        case .lowercased: { $0?.setKeyboardType(.alphabetic(.uppercased)) }
+        default: nil
+        }
+    }
+    
+    var standardReleaseAction: KeyboardAction.GestureAction? {
+        switch self {
+        case .lowercased: nil
+        default: { $0?.setKeyboardType(.alphabetic(.lowercased)) }
+        }
+    }
+}